requirements.txt
openai==1.11.0
Pillow
python-dotenv==1.0.0
rich==13.4.2
numpy==1.24.3
tiktoken==0.5.2
psutil==5.9.8
tqdm==4.66.2
tabulate==0.9.0
requests==2.31.0 